[WIP] - Refactor data processing and repetitions calculations
A new functions to update the chart has been added to remove this functionality from the data treatment function and the post-processing function has been split in two, a process one and an evaluation one. The former fixes and clean the data and the latter calculates repetitions based on the data.