[WIP] - iOS app properly displays watchOS app data
Instead of delivering every motion update as it's generated, due to lag issue, we store them to send the whole dataset at once. To avoid transfer limit between the watch and the phone, we need to use the transferFile function, saving the dataset on a temporal file. When the dataset is received on the phone, every motion data is processed as intended and the chart displays the whole set.